requirements.txt
PyQt6
pytesseract
openai
pynput
Pillow
python-dotenv